// front-end/types/cms.type.ts
import z from "zod";

const CmsReqSchama = z.object({
  id: z.string(),
  banner_image: z.string().optional(),
  bannermobile: z.string().optional(),
  logo_image: z.string().optional(),
  section1_header: z.string().optional(),
  section1_sub: z.string().optional(),
  section2_header: z.string().optional(),
  section2_sub: z.string().optional(),
  section2_banner_url: z.string().optional(),
  section2_sub2: z.string().optional(),
  section3_header: z.string().optional(),
  section3_sub: z.string().optional(),
  section3_item1_image: z.string().optional(),
  section3_item1_header: z.string().optional(),
  section3_item1_sub: z.string().optional(),
  section3_item2_image: z.string().optional(),
  section3_item2_header: z.string().optional(),
  section3_item2_sub: z.string().optional(),
  section3_item3_image: z.string().optional(),
  section3_item3_header: z.string().optional(),
  section3_item3_sub: z.string().optional(),
  promotion_head: z.string().optional(),
  promotion_banner: z.string().optional(),
  promotion_item_image: z.string().optional(),
  promotion_detail: z.string().optional(),
  consult_head: z.string().optional(),
  consult_detail: z.string().optional(),
  service_head: z.string().optional(),
  service_detail: z.string().optional(),
  icon_service_one: z.string().optional(),
  service_item_head_one: z.string().optional(),
  service_item_detail_one: z.string().optional(),
  icon_service_two: z.string().optional(),
  service_item_head_two: z.string().optional(),
  service_item_detail_two: z.string().optional(),
  icon_service_three: z.string().optional(),
  service_item_head_three: z.string().optional(),
  service_item_detail_three: z.string().optional(),
  about_head: z.string().optional(),
  about_detail: z.string().optional(),
  about_image: z.string().optional(),
});

export type Cms = z.infer<typeof CmsReqSchama>;